src/app/pages/products/products.component.spec.ts
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ComponentFixture, TestBed } from '@angular/core/testing';

import { ProductsComponent } from './products.component';

describe('ProductsComponent', () => {
  let component: ProductsComponent;
  let fixture: ComponentFixture<ProductsComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [ HttpClientTestingModule ],
      declarations: [ ProductsComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(ProductsComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });

  it('should create the products component', () => {
    const fixture = TestBed.createComponent(ProductsComponent);
    const products = fixture.componentInstance;
    expect(products).toBeTruthy();
  });

  it('should render the products component', () => {
    const compiled = fixture.nativeElement;
    expect(compiled.querySelector('div')).toBeTruthy();
  });
});
